+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* **Refactoring** Refac the carousel to use @vtex/slick-components.
+* **Import** Refac the carousel to use @vtex/slick-components.
 
 ## [0.2.1] - 2018-18-19